Get log files written out correctly for multiple TDP (#492)

+++ drizzlepac/hlautils/processing_utils.py
@@ -79,7 +79,7 @@ def refine_product_headers(product, total_obj_list):
     phdu['date-beg'] = (Time(phdu['expstart'], format='mjd').iso, "Starting Date and Time")
     phdu['date-end'] = (Time(phdu['expend'], format='mjd').iso, "Ending Date and Time")
 
-    phdu['equinox'] = hdu[('sci',1)].header['equinox'] if 'equinox' in hdu[('sci',1)].header else 2000.0
+    phdu['equinox'] = hdu[('sci', 1)].header['equinox'] if 'equinox' in hdu[('sci', 1)].header else 2000.0
 
     # Re-format ACS filter specification
     if phdu['instrume'] == 'ACS':
@@ -227,9 +227,14 @@ def _process_input(input):
             raise ValueError
 
     return hdu, closefits
-    
-def appendTrlFile(trlfile, drizfile):
+
+def append_trl_file(trlfile, drizfile, clean=True):
     """ Append log file to already existing log or trailer file.
+
+    Parameters
+    -----------
+    clean : bool
+        Remove the `drizfile` or not when finished appending it to `trlfile`
     """
     if not os.path.exists(drizfile):
         return
@@ -248,6 +253,6 @@ def appendTrlFile(trlfile, drizfile):
     ftrl.close()
     fdriz.close()
 
-    # Now, clean up astrodrizzle trailer file
-    os.remove(drizfile)
-
+    if clean:
+        # Now, clean up astrodrizzle trailer file
+        os.remove(drizfile)
